refactor: change all helper functions path
