+++ src/app/shared/components/navbar/navbar.component.ts
@@ -1,7 +1,7 @@
 import { Component, inject, OnInit, output, signal } from '@angular/core';
 import { ProfileCardComponent } from './components/profile-card/profile-card.component';
 import { BrandComponent } from '../brand/brand.component';
-import { setPageHeader } from '../../helpers/constants.helper';
+import { setPageHeader } from '../../helpers/functions.helper';
 import { NavigationEnd, Router } from '@angular/router';
 import { HeaderTitleComponent } from '../header-title/header-title.component';
